refactor: organize modules into package structure
