--- __init__.py
@@ -1,1 +1,0 @@
-from zfn_api import Client
