--- requirements.txt
@@ -1,3 +1,0 @@
-pyquery==1.4.3
-requests==2.27.1
-rsa==4.8
