+++ zfn_api/constants.py
@@ -1,0 +1,15 @@
+RASPIANIE = [
+    ["8:00", "8:40"],
+    ["8:45", "9:25"],
+    ["9:30", "10:10"],
+    ["10:30", "11:10"],
+    ["11:15", "11:55"],
+    ["14:30", "15:10"],
+    ["15:15", "15:55"],
+    ["16:05", "16:45"],
+    ["16:50", "17:30"],
+    ["18:40", "19:20"],
+    ["19:25", "20:05"],
+    ["20:10", "20:50"],
+    ["20:55", "21:35"],
+]
